Rename comments-data to comments-slice
